+++ config/config.js
@@ -2,10 +2,10 @@ const config = {
   botName: process.env.botName,
   PORT: process.env.PORT,
   mongoURI: process.env.mongoURI,
-  sessionSecret: process.env.sessionSecret,
-  local_db: process.env.local_db,
-  db_name: process.env.db_name,
-  sessionMaxAge: process.env.sessionMaxAge,
+  SESSION_SECRET_KEY: process.env.SESSION_SECRET_KEY,
+  MONGODB_URL: process.env.MONGODB_URL,
+  DB_NAME: process.env.DB_NAME,
+  SESSION_MAX_AGE: process.env.SESSION_MAX_AGE,
 };
 
 const cors = {
